# docs/swagger/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "phpList4 REST API",
        "version": "4.0.0"
    },
    "paths": {
        "/api/v2/lists": {
            "get": {
                "tags": [
                    "lists"
                ],
                "summary": "Gets a list of all subscriber lists.",
                "description": "Returns a json list of all subscriber lists.",
                "operationId": "PhpList\\RestBundle\\Controller\\ListController::cgetAction",
                "parameters": [
                    {
                        "name": "session",
                        "in": "header",
                        "description": "Session ID obtained from authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "name": "News",
                                        "description": "News (and some fun stuff)",
                                        "creation_date": "2016-06-22T15:01:17+00:00",
                                        "list_position": 12,
                                        "subject_prefix": "phpList",
                                        "public": true,
                                        "category": "news",
                                        "id": 1
                                    },
                                    {
                                        "name": "More news",
                                        "description": "",
                                        "creation_date": "2016-06-22T15:01:17+00:00",
                                        "list_position": 12,
                                        "subject_prefix": "",
                                        "public": true,
                                        "category": "",
                                        "id": 2
                                    }
                                ]
                            }
                        }
                    },
                    "403": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Gets a list of all subscriber lists.",
                                            "type": "string",
                                            "example": "No valid session key was provided as basic auth password."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/lists/{list}": {
            "get": {
                "tags": [
                    "lists"
                ],
                "summary": "Gets a subscriber list.",
                "description": "Returns a single subscriber lists with specified ID",
                "operationId": "PhpList\\RestBundle\\Controller\\ListController::getAction",
                "parameters": [
                    {
                        "name": "list",
                        "in": "path",
                        "description": "List ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "session",
                        "in": "header",
                        "description": "Session ID obtained from authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "name": "News",
                                        "description": "News (and some fun stuff)",
                                        "creation_date": "2016-06-22T15:01:17+00:00",
                                        "list_position": 12,
                                        "subject_prefix": "phpList",
                                        "public": true,
                                        "category": "news",
                                        "id": 1
                                    }
                                ]
                            }
                        }
                    },
                    "403": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Gets a subscriber list.",
                                            "type": "string",
                                            "example": "No valid session key was provided as basic auth password."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Gets a subscriber list.",
                                            "type": "string",
                                            "example": "There is no list with that ID."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "lists"
                ],
                "summary": "Deletes a list.",
                "description": "Deletes a single subscriber list passed as",
                "operationId": "PhpList\\RestBundle\\Controller\\ListController::deleteAction",
                "parameters": [
                    {
                        "name": "session",
                        "in": "header",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Deletes a subscriber list.",
                                            "type": "string",
                                            "example": "No valid session key was provided as basic auth password or You do not have access to this session."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Deletes a subscriber list.",
                                            "type": "string",
                                            "example": "There is no session with that ID."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/lists/{list}/members": {
            "get": {
                "tags": [
                    "lists"
                ],
                "summary": "Gets a list of all subscribers (members) of a subscriber list.",
                "description": "Returns a json list of all subscriber lists.",
                "operationId": "PhpList\\RestBundle\\Controller\\ListController::getMembersAction",
                "parameters": [
                    {
                        "name": "session",
                        "in": "header",
                        "description": "Session ID obtained from authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "list",
                        "in": "path",
                        "description": "List ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "creation_date": "2016-07-22T15:01:17+00:00",
                                        "email": "oliver@example.com",
                                        "confirmed": true,
                                        "blacklisted": true,
                                        "bounce_count": 17,
                                        "unique_id": "95feb7fe7e06e6c11ca8d0c48cb46e89",
                                        "html_email": true,
                                        "disabled": true,
                                        "id": 1
                                    },
                                    {
                                        "creation_date": "2017-07-22T15:12:17+00:00",
                                        "email": "sam@example.com",
                                        "confirmed": true,
                                        "blacklisted": false,
                                        "bounce_count": 1,
                                        "unique_id": "95feb7fe7e06e6c11ca8d0c48cb4616d",
                                        "html_email": false,
                                        "disabled": false,
                                        "id": 2
                                    }
                                ]
                            }
                        }
                    },
                    "403": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Gets a list of all subscribers (members) of a subscriber list.",
                                            "type": "string",
                                            "example": "No valid session key was provided as basic auth password."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/lists/{list}/count": {
            "get": {
                "tags": [
                    "lists"
                ],
                "summary": "Gets the total number of subscribers of a list",
                "description": "Returns a count of all subscribers in a given list.",
                "operationId": "PhpList\\RestBundle\\Controller\\ListController::getSubscribersCountAction",
                "parameters": [
                    {
                        "name": "session",
                        "in": "header",
                        "description": "Session ID obtained from authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "list",
                        "in": "path",
                        "description": "List ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Gets the total number of subscribers of a list.",
                                            "type": "string",
                                            "example": "No valid session key was provided as basic auth password."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v2/sessions/{session}": {
            "delete": {
                "tags": [
                    "sessions"
                ],
                "summary": "Delete a session.",
                "description": "Delete the session passed as paramater",
                "operationId": "PhpList\\RestBundle\\Controller\\SessionController::deleteAction",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Deletes a session.",
                                            "type": "string",
                                            "example": "No valid session key was provided as basic auth password or You do not have access to this session."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Deletes a session.",
                                            "type": "string",
                                            "example": "There is no session with that ID."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}